Add trait StarkHash
