+++ crates/starknet-types-core/Cargo.toml
@@ -21,7 +21,7 @@ arbitrary = { version = "1.3.0", optional = true, default-features = false }
 num-traits = { version = "0.2.16", default-features = false }
 
 [features]
-default = ["std", "serde", "curve"]
+default = ["std", "serde", "curve", "hash"]
 curve = []
 hash = ["dep:lambdaworks-crypto"]
 std = []
