+++ crates/starknet-types-core/src/hash/mod.rs
@@ -1,0 +1,2 @@
+pub mod pedersen;
+pub mod traits;
